Extract folder browser rename workflow
